build: add Plan 9 mkfile
This mkfile builds the Lua core and most of the libraries, packaging the
results in a library intended for static linking with host programs.

I haven't ported the standard lua.c standalone interpreter because it
doesn't make much sense on Plan 9 due to the lack of dynamic linking
facilities.

A standalone Lua interpreter taking full advantage of Plan 9 interfaces
and packed with useful libraries is provided as a separate project.
It also doubles as a usage example for the library port.

Additionally, the Lua 'os' library build is disabled, since Plan 9
doesn't provide the required libc interfaces.  A fully rewritten
compatible os library might be included in this port in the future, but
for now it lives in the above mentioned Plan 9 interpreter.