los9lib: make os.date accept and output Lua rather than C strings
Lua wants the format string and formatted output to be Lua rather than C
strings, allowing embedded zero bytes to be formatted.
This means we can't use the fmtprint(2) routines to directly construct
the output, so we split the formatting into pieces and use a lua_Buffer
to construct the result Lua string.

This also fixes the '%%' escape.