shim: incorporate shim code into liblua
Having shim code in a separate repository seemed like a good
idea since it could have been reused by different projects.
However, this didn't really occur since I realized that it's
best to provide only the shim code that's needed on a per-port
basis -- for which a simple copy-paste approach works fine.

This change simplifies the project structure and makes it easier
for others to use liblua -- all they need is liblua.

However, it is still neccessary to tell the compiler where
the shim headers are.  It's not great, but better than dumping
all this stuff into lua's own directory.