p9: rewrite everything in C
This is done to minimize the dependencies a library user must
think about if they want to use it in their own programs.
With everything in a single library archive all that needs to
be done is link or load it.

During the conversion the API of several functions got changed
and some new functions got added as well.

If you have installed this library previously you will want to
remove the Lua module path files that got installed in
	/sys/lib/lua/5.4/p9
It doesn't hurt to keep them since this library is preloaded
but someone might get confused in the future.

A rebuild of lu9 and any of your programs will obviously be
needed as well.