shim: fix size_t and ptrdiff_t definitions
Prior to a recent change to 9front which made the usize type
64-bit on 64-bit machines there was no nice way to define a
proper size_t, so I opted to hardcode it to uvlong, which
avoided the need for patching a bunch of tricky upstream code
dealing with size limits and whatnot. This, of course, made
it so that Lua would only compile on 64-bit platforms.

With this change Lua now builds for all the usually supported
9front platforms: amd64, 386, arm, arm64.